[LW-17] Embed Log.* into Wlog.*, merge functionality
