+++ src/System/Wlog/Handler.hs
@@ -1,6 +1,6 @@
-{-# LANGUAGE ConstrainedClassMethods #-}
-{-# LANGUAGE FlexibleContexts        #-}
-{-# LANGUAGE TypeFamilies            #-}
+{-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE TypeFamilies     #-}
+
 {- |
    Module     : System.Log.Handler
    Copyright  : Copyright (C) 2004-2011 John Goerzen
@@ -12,30 +12,24 @@
 
 Definition of log handler support
 
-For some handlers, check out "System.Log.Handler.Simple" and
-"System.Log.Handler.Syslog".
+For some handlers, check out "System.WLog.Handler.Simple" and
+"System.WLog.Handler.Syslog".
 
-Please see "System.Log.Logger" for extensive documentation on the
+Please see "System.WLog.Logger" for extensive documentation on the
 logging system.
 
 Written by John Goerzen, jgoerzen\@complete.org
 -}
-
-module System.Log.Handler
+module System.Wlog.Handler
        ( -- * Basic Types
          LogHandler(..)
        ) where
 
-import           Data.Typeable
-import           System.IO
-import           System.Log
-import           System.Log.Formatter
+import           System.Wlog.Formatter (LogFormatter, nullFormatter)
+import           System.Wlog.Severity  (LogRecord, Severity)
 
-{- | All log handlers should adhere to this. -}
-
-{- | This is the base class for the various log handlers.  They should
-all adhere to this class. -}
-
+-- | This is the base class for the various log handlers.  They should
+-- all adhere to this class.
 class LogHandler a where
     -- TODO it should be data/type family like in 'System.Exception'
     -- | Tag of handler. Is arbitrary. Made for identication.
@@ -43,24 +37,24 @@ class LogHandler a where
 
     -- | Sets the log level. 'handle' will drop items beneath this
     -- level.
-    setLevel :: a -> Priority -> a
+    setLevel :: a -> Severity -> a
 
     -- | Gets the current level.
-    getLevel :: a -> Priority
+    getLevel :: a -> Severity
 
     -- | Set a log formatter to customize the log format for this Handler
     setFormatter :: a -> LogFormatter a -> a
     getFormatter :: a -> LogFormatter a
-    getFormatter h = nullFormatter
+    getFormatter _h = nullFormatter
 
     -- | Logs an event if it meets the requirements
     -- given by the most recent call to 'setLevel'.
     handle :: a -> LogRecord -> String -> IO ()
     handle h (pri, msg) logname =
         if pri >= (getLevel h)
-           then do formattedMsg <- (getFormatter h) h (pri,msg) logname
-                   emit h (pri, formattedMsg) logname
-           else return ()
+        then do formattedMsg <- (getFormatter h) h (pri,msg) logname
+                emit h (pri, formattedMsg) logname
+        else return ()
 
     -- | Forces an event to be logged regardless of
     -- the configured level.
